Change the location of example
