+++ scala-package/examples/src/main/scala/org/apache/mxnetexamples/imclassification/README.md
